Extracted Arena game briefing from TutorialPane to separate ArenaBriefingPane
